+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changes to this project are documented in this file. More detail and links can b
 - AUTOTOOL-1051: Event Listener unable to classify AFM DoS event
 - AUTOTOOL-1037: Splunk legacy tmstats - include last_cycle_count
 - AUTOTOOL-1019: Splunk legacy tmstats - add tenant and application data
+- AUTOTOOL-1128: Declarations with large secrets may timeout
 ### Changed
 ### Removed